moved api utils.js under src/functions
