+++ api/package.json
@@ -5,11 +5,20 @@
   "main": "index.js",
   "scripts": {
     "start": "func start",
-    "build": "echo 'No build step required for API'"
+    "build": "npm run generate-function-json",
+    "generate-function-json": "tsc",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "test:watch": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watch",
+    "test:coverage": "node --experimental-vm-modules node_modules/jest/bin/jest.js --coverage"
   },
   "dependencies": {
-    "@azure/functions": "^4.0.0",
+    "@azure/functions": "^4.7.0",
     "axios": "^1.8.4",
     "azure-function-express": "^1.1.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.14",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.3.1"
   }
 }
